+++ packages/services/src/constants.ts
@@ -1,5 +1,5 @@
 import { TBaseConfig } from '@etransfer/request';
-import { TAuthApiBaseParams } from '@etransfer/types';
+import { PortkeyVersion, TAuthApiBaseParams, TOtherChainAuthApiBaseParams } from '@etransfer/types';
 
 export const DEFAULT_METHOD = 'GET';
 
@@ -30,6 +30,19 @@ const WithdrawApiList: Record<string, TBaseConfig> = {
   },
 };
 
+const TransferApiList: Record<string, TBaseConfig> = {
+  getTokenNetworkRelation: '/api/etransfer/network/tokens',
+  getTransferInfo: '/api/etransfer/transfer/info',
+  createTransferOrder: {
+    target: '/api/etransfer/transfer/order',
+    baseConfig: { method: 'POST' },
+  },
+  updateTransferOrder: {
+    target: '/api/etransfer/transfer',
+    baseConfig: { method: 'PUT' },
+  },
+};
+
 const HistoryApiList: Record<string, TBaseConfig> = {
   getRecordsList: '/api/etransfer/record/list',
   getRecordStatus: '/api/etransfer/record/status',
@@ -38,6 +51,7 @@ const HistoryApiList: Record<string, TBaseConfig> = {
 
 const UserApiList: Record<string, TBaseConfig> = {
   checkEOARegistration: '/api/etransfer/user/check-eoa-registration',
+  checkRegistration: '/api/etransfer/user/check-registration',
 };
 
 export const API_LIST: Record<string, Record<string, TBaseConfig>> = {
@@ -45,6 +59,7 @@ export const API_LIST: Record<string, Record<string, TBaseConfig>> = {
   common: CommonApiList,
   deposit: DepositApiList,
   withdraw: WithdrawApiList,
+  transfer: TransferApiList,
   records: HistoryApiList,
   user: UserApiList,
 };
@@ -53,10 +68,19 @@ export enum CANCEL_TOKEN_SOURCE_KEY {
   GET_DEPOSIT_INFO = 'GET_DEPOSIT_INFO',
   GET_WITHDRAW_INFO = 'GET_WITHDRAW_INFO',
   GET_NETWORK_LIST = 'GET_NETWORK_LIST',
+  GET_TRANSFER_INFO = 'GET_TRANSFER_INFO',
 }
 
 export const AUTH_API_BASE_PARAMS: TAuthApiBaseParams = {
   grant_type: 'signature',
   scope: 'ETransferServer',
   client_id: 'ETransferServer_App',
+};
+
+export const OTHER_CHAIN_AUTH_API_BASE_PARAMS: TOtherChainAuthApiBaseParams = {
+  grant_type: 'signature',
+  scope: 'ETransferServer',
+  client_id: 'ETransferServer_App',
+  version: PortkeyVersion.v2,
+  source: 'wallet',
 };