+++ packages/types/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.13.0](https://github.com/ETransferProject/etransfer-toolkit/compare/v1.12.2-alpha.1...v1.13.0) (2025-02-10)
+
+### Features
+
+- add other chain auth token api ([e5f4d46](https://github.com/ETransferProject/etransfer-toolkit/commit/e5f4d4601de8ba2bc76c75e651829ce3ea94bdaf))
+- add transfer services ([3401208](https://github.com/ETransferProject/etransfer-toolkit/commit/34012081339b487ad16953443bd0b54cf9ab94e4))
+
 ## [1.12.2-alpha.1](https://github.com/ETransferProject/etransfer-toolkit/compare/v1.12.2-alpha.0...v1.12.2-alpha.1) (2025-01-08)
 
 **Note:** Version bump only for package @etransfer/types